requirements.txt
asgiref==3.2.7
autopep8==1.5
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.2.1
django-rest-auth==0.9.5
django-smtp-ssl==1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==1.4.6
packaging==20.3
Pillow==7.1.1
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
